Add files via upload
This  calculator, built with Python and Tkinter, offers both basic and scientific modes for versatile calculations. It features theme toggling between light and dark modes, a history log to track past calculations, and a responsive UI that adjusts to various screen sizes. To get started, clone the repository, install the dependencies, and run the application